% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Returns hello world string.}
\usage{
hello()
}
\value{
"Hello, world!"
}
\description{
Returns hello world string.
}
\examples{
hello()
hello_string <- hello()
print(paste0("Johny says, '", hello_string, "'"))
}
